requirements.txt
GitPython==3.1.29
librosa==0.8.1
matplotlib==3.1.3
numpy==1.21.0
pdoc==12.2.2
scipy==1.4.1
setuptools==50.3.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cu113
--find-links https://download.pytorch.org/whl/torch_stable.html
torchaudio==0.11.0+cu113
tqdm==4.50.2
wandb==0.13.5